Completed "Add Mentor" button functionality
The add mentor button now functions properly. Additional mentors are
hidden upon loading the webpage, and clicking the button will display
additional mentor information up to four mentors.

If the user attempts to add additional mentors past the limit of four,
then the webpage sends a javascript alert informing them of the limit.
If anyone knows of a better way to do this than with an alert box,
please feel free to update it.

There was also a problem with the "Add Mentor" button submitting the
form upon clicking. I fixed this by changing its type to "button". It
seemed a little redundant, but apparently it cleared up the problem.