Completed functionality for pricing section
There are currently three functions for the pricing section. A function
is called when a numerical value is either added or subtracted from the
quantity boxes for comprehensive, algII, or geometry. The functions
check to see if the quantity value is either increasing or decreasing,
and then they will update the price and grand total box.

Note: Each of the functions are identical except for the Jquery
selectors. If there is a way to make the selector variable, then we
could cut the pricing code down to about 1/3 its current size. It works
for now, but if anyone wants to update it, then feel free.